GPT LLM now takes images. We removed Rekognition label
